perf: experimental changes to improve resolve performance
