+++ generator/web/src/main/kotlin/me/kcra/takenaka/generator/web/WebGenerator.kt
@@ -26,6 +26,7 @@ import kotlinx.html.dom.serialize
 import me.kcra.takenaka.core.CompositeWorkspace
 import me.kcra.takenaka.core.Workspace
 import me.kcra.takenaka.core.mapping.ElementRemapper
+import me.kcra.takenaka.core.mapping.MappingsMap
 import me.kcra.takenaka.core.mapping.adapter.completeInnerClassNames
 import me.kcra.takenaka.core.mapping.adapter.replaceCraftBukkitNMSVersion
 import me.kcra.takenaka.core.mapping.ancestry.classAncestryTreeOf
@@ -33,7 +34,7 @@ import me.kcra.takenaka.core.mapping.allNamespaceIds
 import me.kcra.takenaka.core.mapping.resolve.VanillaMappingContributor
 import me.kcra.takenaka.core.mapping.resolve.modifiers
 import me.kcra.takenaka.core.util.objectMapper
-import me.kcra.takenaka.generator.common.AbstractGenerator
+import me.kcra.takenaka.generator.common.AbstractResolvingGenerator
 import me.kcra.takenaka.generator.common.ContributorProvider
 import me.kcra.takenaka.generator.web.components.footerComponent
 import me.kcra.takenaka.generator.web.components.navComponent
@@ -84,7 +85,7 @@ class WebGenerator(
     val index: ClassSearchIndex = emptyClassSearchIndex(),
     val spigotLikeNamespaces: List<String> = emptyList(),
     val historyAllowedNamespaces: List<String> = namespaceFriendlinessIndex - MappingUtil.NS_SOURCE_FALLBACK
-) : AbstractGenerator(
+) : AbstractResolvingGenerator(
     workspace,
     versions,
     mappingWorkspace,
@@ -119,9 +120,9 @@ class WebGenerator(
     val namespacesByFriendlyNames = namespaces.mapKeys { it.value.friendlyName }
 
     /**
-     * Launches the generator.
+     * Launches the generator with a pre-determined set of mappings.
      */
-    override suspend fun generate() {
+    override suspend fun generate(mappings: MappingsMap) {
         val styleConsumer = DefaultStyleConsumer()
 
         generationContext(styleConsumer = styleConsumer::apply) {
